+++ docs/source/usage.rst
@@ -1,9 +1,11 @@
 Usage
 =====
 
-.. autoclass:: oncvpsp_tools.ONCVPSPInput
+`oncvpsp-tools` implements the contents of `oncvpsp.x` input and output files as Python dataclasses.
+
+.. autoclass:: oncvpsp_tools.input.ONCVPSPInput
     :members:
 
-.. autoclass:: oncvpsp_tools.ONCVPSPOutput
+.. autoclass:: oncvpsp_tools.output.ONCVPSPOutput
     :members:
