+++ setup.cfg
@@ -84,7 +84,7 @@ tests =
     pytest
     coverage
 docs =
-    sphinx>=7.2.6
+    sphinx
     sphinx-rtd-theme
     sphinx-click
     sphinx_automodapi
@@ -152,6 +152,7 @@ ignore =
     W503 # Line break before binary operator (flake8 is wrong)
     E203 # whitespace before ':'
     E741 # one-character variables
+    I005 # missing import (prompted by use of noqa)
 exclude =
     .tox,
     .git,
