+++ src/oncvpsp_tools/__init__.py
@@ -3,4 +3,5 @@
 """Tools for handling input and output files of oncvpsp.x."""
 
 from .api import *  # noqa
-from .oncvpsp import *  # noqa
+from .input import ONCVPSPInput  # noqa
+from .output import ONCVPSPOutput  # noqa
